python/requirements.txt
numpy==1.23.4
paho-mqtt==1.6.1
pandas==1.5.1
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
